Mild cleanup/refactor of QPlotWidget
